cleans folder structure
