+++ packages/core/src/components/Panel/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./Panel";
+export * from "./Section";
