--- packages/core/src/components/PanelSection/example.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="bim-ui-dark">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Panel Section</title>
-  <style>
-    body {
-      margin: 0;
-    }
-
-    html {
-      font-family: sans-serif;
-    }
-
-    bim-panel-section {
-      position: absolute;
-      top: 1rem;
-      left: 1rem;
-      background-color: var(--bim-ui_bg-base);
-      width: 20rem;
-      border-radius: 0.375rem;
-    }
-  </style>
-</head>
-
-<body>
-  <bim-panel-section label="Your Panel Section" icon="solar:settings-bold">
-    <bim-color-input label="Color"></bim-color-input>
-  </bim-panel-section>
-  <script type="module" src="./example.ts"></script>
-</body>
-
-</html>
