--- packages/core/src/components/PanelSection/example.ts
@@ -1,3 +1,0 @@
-import * as BUI from "../..";
-
-BUI.Manager.init();
